Add TagsCoverter.
