+++ .gitignore
@@ -240,3 +240,4 @@ _Pvt_Extensions
 **/.vscode
 **/node_modules
 /cs/TestSolution
+/cs/Markdown/HTML.cs
